add `ui/scrubber-preview` tests
